Move class changes to correct location
