+++ addons/Redd_Main/CfgAmmo.hpp
@@ -3,4 +3,10 @@ class CfgAmmo {
     class Redd_Milan_AT: M_Titan_AT {
         maxControlRange = 1975;
     };
+
+    class B_30mm_HE_Tracer_Red;
+    class Redd_B_30mm_HE_Tracer_Red: B_30mm_HE_Tracer_Red {
+        indirectHit = 7;
+        indirectHitRange = 2.5;
+    };
 };
