+++ addons/Redd_Main/CfgMagazines.hpp
@@ -1,16 +1,13 @@
-class CfgMagazines
-{
+class CfgMagazines {
     class 140Rnd_30mm_MP_shells_Tracer_Red;
-    class Redd_MK20_HE_Mag: 140Rnd_30mm_MP_shells_Tracer_Red
-    {
+    class Redd_MK20_HE_Mag: 140Rnd_30mm_MP_shells_Tracer_Red {
         displayName = "20mm HE-T";
         displayNameShort = "HE-T";
         count = 275;
     };
 
     class 60Rnd_30mm_APFSDS_shells_Tracer_Red;
-    class Redd_MK20_AP_Mag: 60Rnd_30mm_APFSDS_shells_Tracer_Red
-    {
+    class Redd_MK20_AP_Mag: 60Rnd_30mm_APFSDS_shells_Tracer_Red {
         displayName = "20mm APDS-T";
         displayNameShort = "APDS-T";
         count = 125;
