--- addons/Redd_Marder/CfgAmmo.hpp
@@ -1,10 +1,0 @@
-class CfgAmmo
-{
-    class B_30mm_HE_Tracer_Red;
-
-    class Redd_B_30mm_HE_Tracer_Red: B_30mm_HE_Tracer_Red
-    {
-        indirectHit = 7;
-        indirectHitRange = 2.5;
-    };
-};
