--- addons/Redd_Marder/CfgWeapons.hpp
@@ -1,7 +1,0 @@
-class CfgWeapons {
-    class Redd_SmokeLauncher;
-    class GVAR(SmokeLauncher): Redd_SmokeLauncher {
-        magazines[] = {"Redd_SmokeLauncherMag"};
-        showToPlayer = 0;
-    };
-};
