+++ addons/Redd_Marder/config.cpp
@@ -24,8 +24,5 @@ class CfgPatches {
     };
 };
 
-#include "CfgAmmo.hpp"
 #include "CfgEventHandlers.hpp"
-#include "CfgMagazines.hpp"
 #include "CfgVehicles.hpp"
-#include "CfgWeapons.hpp"
